tests separated in test folder, import doesn´t work -> setup.py has to import something i guess
